fix: fixing some errors related to acess id before enities are persist

+++ src/main/java/com/example/projetoApiVendasEmSpring/dtos/financialTransaction/FinancialTransactionOutputDto.java
@@ -22,8 +22,8 @@ public record FinancialTransactionOutputDto(
           FinancialPaymentMethod paymentMethod,
           FinancialPaymentTerm paymentTerm,
           List<InstallmentOutputDto> installments,
-          int installmentCount,
-          int paidInstallmentCount,
+          long installmentCount,
+          long paidInstallmentCount,
           BigDecimal totalPaidAmount
 ) {
 }
